+++ toolkit/cli/smart-contracts-commands/src/reserve.rs
@@ -4,6 +4,7 @@ use partner_chains_cardano_offchain::{
 	await_tx::FixedDelayRetries,
 	reserve::{
 		create::{create_reserve_utxo, ReserveParameters},
+		deposit::{deposit_to_reserve, TokenAmount},
 		init::init_reserve_management,
 	},
 };
@@ -15,6 +16,7 @@ pub enum ReserveCmd {
 	/// Initialize the reserve management system for your chain
 	Init(InitReserveCmd),
 	Create(CreateReserveCmd),
+	Deposit(DepositReserveCmd),
 }
 
 impl ReserveCmd {
@@ -22,6 +24,7 @@ impl ReserveCmd {
 		match self {
 			Self::Init(cmd) => cmd.execute().await,
 			Self::Create(cmd) => cmd.execute().await,
+			Self::Deposit(cmd) => cmd.execute().await,
 		}
 	}
 }
@@ -86,6 +89,39 @@ impl CreateReserveCmd {
 				token: self.token,
 				initial_deposit: self.initial_deposit_amount,
 			},
+			self.genesis_utxo,
+			payment_key.0,
+			&ogmios_client,
+			&FixedDelayRetries::two_minutes(),
+		)
+		.await?;
+		Ok(())
+	}
+}
+
+#[derive(Clone, Debug, clap::Parser)]
+pub struct DepositReserveCmd {
+	#[clap(flatten)]
+	common_arguments: crate::CommonArguments,
+	#[clap(flatten)]
+	payment_key_file: PaymentFilePath,
+	/// Genesis UTXO of the partner-chain.
+	#[arg(long, short('c'))]
+	genesis_utxo: UtxoId,
+	/// Use either "Ada" or encoded asset id in form <policy_id_hex>.<asset_name_hex>.
+	#[arg(long)]
+	token: TokenId,
+	/// Amount of tokens to deposit. They must be present in the payment wallet.
+	#[arg(long)]
+	amount: u64,
+}
+
+impl DepositReserveCmd {
+	pub async fn execute(self) -> crate::CmdResult<()> {
+		let payment_key = self.payment_key_file.read_key()?;
+		let ogmios_client = client_for_url(&self.common_arguments.ogmios_url).await?;
+		let _ = deposit_to_reserve(
+			TokenAmount { token: self.token, amount: self.amount },
 			self.genesis_utxo,
 			payment_key.0,
 			&ogmios_client,